+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetdatapointsop.py
@@ -600,7 +600,7 @@ class ListDatasetDatapointsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-06-03T09:49:54.847Z")
+    updated: Optional[datetime] = parse_datetime("2025-06-03T12:31:41.205Z")
     r"""The date and time the resource was last updated"""