+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JWTK5AG51JNNRH3N94HFPEH4"
+        "01JWTWDH15M8K1GE34DZACKDM8"
     )
     r"""The id of the resource"""