+++ packages/orq-rc/RELEASES.md
@@ -1649,3 +1649,13 @@ Based on:
 - [python v3.7.0-rc.7] packages/orq-rc
 ### Releases
 - [PyPI v3.7.0-rc.7] https://pypi.org/project/orq-ai-sdk/3.7.0-rc.7 - packages/orq-rc
+
+## 2025-06-03 12:33:09
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.555.0 (2.618.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.7.0-rc.8] packages/orq-rc
+### Releases
+- [PyPI v3.7.0-rc.8] https://pypi.org/project/orq-ai-sdk/3.7.0-rc.8 - packages/orq-rc